Rename node_depth polars file and move helper into it
Rename _alifestd_mark_node_depth_asexual_polars to
_alifestd_mark_node_depth_polars and move
_alifestd_calc_node_depth_asexual_contiguous helper into
the polars file, following the colless-like index pattern
where shared helpers are co-located with the implementation.
