+++ hstrat/_auxiliary_lib/_alifestd_calc_mrca_id_vector_asexual_polars.py
@@ -13,7 +13,7 @@ from ._alifestd_has_contiguous_ids_polars import (
 from ._alifestd_is_topologically_sorted_polars import (
     alifestd_is_topologically_sorted_polars,
 )
-from ._alifestd_mark_node_depth_asexual import (
+from ._alifestd_mark_node_depth_polars import (
     _alifestd_calc_node_depth_asexual_contiguous,
 )
 from ._alifestd_try_add_ancestor_id_col_polars import (
