+++ hstrat/_auxiliary_lib/_alifestd_mark_node_depth_polars.py
@@ -4,6 +4,7 @@ import os
 
 import joinem
 from joinem._dataframe_cli import _add_parser_base, _run_dataframe_cli
+import numpy as np
 import polars as pl
 
 from ._alifestd_has_contiguous_ids_polars import (
@@ -12,19 +13,33 @@ from ._alifestd_has_contiguous_ids_polars import (
 from ._alifestd_is_topologically_sorted_polars import (
     alifestd_is_topologically_sorted_polars,
 )
-from ._alifestd_mark_node_depth_asexual import (
-    _alifestd_calc_node_depth_asexual_contiguous,
-)
 from ._alifestd_try_add_ancestor_id_col_polars import (
     alifestd_try_add_ancestor_id_col_polars,
 )
 from ._configure_prod_logging import configure_prod_logging
 from ._format_cli_description import format_cli_description
 from ._get_hstrat_version import get_hstrat_version
+from ._jit import jit
 from ._log_context_duration import log_context_duration
 
 
-def alifestd_mark_node_depth_asexual_polars(
+@jit(nopython=True)
+def _alifestd_calc_node_depth_asexual_contiguous(
+    ancestor_ids: np.ndarray,
+) -> np.ndarray:
+    """Optimized implementation for asexual phylogenies with contiguous ids."""
+    ancestor_ids = ancestor_ids.astype(np.uint64)
+    node_depths = np.full_like(ancestor_ids, -1, dtype=np.int64)
+
+    for id_, _ in enumerate(ancestor_ids):
+        ancestor_id = ancestor_ids[id_]
+        ancestor_depth = node_depths[ancestor_id]
+        node_depths[id_] = ancestor_depth + 1
+
+    return node_depths
+
+
+def alifestd_mark_node_depth_polars(
     phylogeny_df: pl.DataFrame,
 ) -> pl.DataFrame:
     """Add column `node_depth`, counting the number of nodes between a
@@ -49,7 +64,7 @@ def alifestd_mark_node_depth_asexual_polars(
         Pandas-based implementation.
     """
     logging.info(
-        "- alifestd_mark_node_depth_asexual_polars: "
+        "- alifestd_mark_node_depth_polars: "
         "adding ancestor_id col...",
     )
     phylogeny_df = alifestd_try_add_ancestor_id_col_polars(phylogeny_df)
@@ -60,7 +75,7 @@ def alifestd_mark_node_depth_asexual_polars(
         )
 
     logging.info(
-        "- alifestd_mark_node_depth_asexual_polars: "
+        "- alifestd_mark_node_depth_polars: "
         "checking contiguous ids...",
     )
     if not alifestd_has_contiguous_ids_polars(phylogeny_df):
@@ -69,7 +84,7 @@ def alifestd_mark_node_depth_asexual_polars(
         )
 
     logging.info(
-        "- alifestd_mark_node_depth_asexual_polars: "
+        "- alifestd_mark_node_depth_polars: "
         "checking topological sort...",
     )
     if not alifestd_is_topologically_sorted_polars(phylogeny_df):
@@ -78,7 +93,7 @@ def alifestd_mark_node_depth_asexual_polars(
         )
 
     logging.info(
-        "- alifestd_mark_node_depth_asexual_polars: "
+        "- alifestd_mark_node_depth_polars: "
         "extracting ancestor ids...",
     )
     ancestor_ids = (
@@ -90,7 +105,7 @@ def alifestd_mark_node_depth_asexual_polars(
     )
 
     logging.info(
-        "- alifestd_mark_node_depth_asexual_polars: "
+        "- alifestd_mark_node_depth_polars: "
         "calculating node depths...",
     )
     node_depths = _alifestd_calc_node_depth_asexual_contiguous(ancestor_ids)
@@ -133,7 +148,7 @@ def _create_parser() -> argparse.ArgumentParser:
     parser = _add_parser_base(
         parser=parser,
         dfcli_module=(
-            "hstrat._auxiliary_lib._alifestd_mark_node_depth_asexual_polars"
+            "hstrat._auxiliary_lib._alifestd_mark_node_depth_polars"
         ),
         dfcli_version=get_hstrat_version(),
     )
@@ -149,12 +164,12 @@ if __name__ == "__main__":
     try:
         with log_context_duration(
             "hstrat._auxiliary_lib"
-            "._alifestd_mark_node_depth_asexual_polars",
+            "._alifestd_mark_node_depth_polars",
             logging.info,
         ):
             _run_dataframe_cli(
                 base_parser=parser,
-                output_dataframe_op=(alifestd_mark_node_depth_asexual_polars),
+                output_dataframe_op=(alifestd_mark_node_depth_polars),
             )
     except NotImplementedError as e:
         logging.error(
