+++ tests/test_hstrat/test_auxiliary_lib/test_alifestd_mark_node_depth_polars_cli.py
@@ -5,38 +5,38 @@ import subprocess
 assets = os.path.join(os.path.dirname(os.path.abspath(__file__)), "assets")
 
 
-def test_alifestd_mark_node_depth_asexual_polars_cli_help():
+def test_alifestd_mark_node_depth_polars_cli_help():
     subprocess.run(  # nosec B603
         [
             "python3",
             "-m",
-            "hstrat._auxiliary_lib._alifestd_mark_node_depth_asexual_polars",
+            "hstrat._auxiliary_lib._alifestd_mark_node_depth_polars",
             "--help",
         ],
         check=True,
     )
 
 
-def test_alifestd_mark_node_depth_asexual_polars_cli_version():
+def test_alifestd_mark_node_depth_polars_cli_version():
     subprocess.run(  # nosec B603
         [
             "python3",
             "-m",
-            "hstrat._auxiliary_lib._alifestd_mark_node_depth_asexual_polars",
+            "hstrat._auxiliary_lib._alifestd_mark_node_depth_polars",
             "--version",
         ],
         check=True,
     )
 
 
-def test_alifestd_mark_node_depth_asexual_polars_cli_csv():
-    output_file = "/tmp/hstrat_alifestd_mark_node_depth_asexual_polars.csv"
+def test_alifestd_mark_node_depth_polars_cli_csv():
+    output_file = "/tmp/hstrat_alifestd_mark_node_depth_polars.csv"
     pathlib.Path(output_file).unlink(missing_ok=True)
     subprocess.run(  # nosec B603
         [
             "python3",
             "-m",
-            "hstrat._auxiliary_lib._alifestd_mark_node_depth_asexual_polars",
+            "hstrat._auxiliary_lib._alifestd_mark_node_depth_polars",
             "--eager-write",
             output_file,
         ],
@@ -46,16 +46,16 @@ def test_alifestd_mark_node_depth_asexual_polars_cli_csv():
     assert os.path.exists(output_file)
 
 
-def test_alifestd_mark_node_depth_asexual_polars_cli_empty():
+def test_alifestd_mark_node_depth_polars_cli_empty():
     output_file = (
-        "/tmp/hstrat_alifestd_mark_node_depth_asexual_polars_empty.csv"
+        "/tmp/hstrat_alifestd_mark_node_depth_polars_empty.csv"
     )
     pathlib.Path(output_file).unlink(missing_ok=True)
     subprocess.run(  # nosec B603
         [
             "python3",
             "-m",
-            "hstrat._auxiliary_lib._alifestd_mark_node_depth_asexual_polars",
+            "hstrat._auxiliary_lib._alifestd_mark_node_depth_polars",
             "--eager-write",
             output_file,
         ],
